make a library not a binary
