+++ .gitignore
@@ -13,3 +13,11 @@ Cargo.lock
 # Added by cargo
 
 /target
+
+
+# Added by cargo
+#
+# already existing elements were commented out
+
+#/target
+/Cargo.lock
